refactor: 지도 공통 레이아웃으로 추출
