+++ src/components/template/Layout.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Footer from './Footer';
-import { Navbar } from './Navbar';
+import MapLayout from '../template/MapLayout/MapLayout.tsx';
 
 interface LayoutProps {
   children: React.ReactNode;
@@ -8,20 +7,9 @@ interface LayoutProps {
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
   return (
-    <div>
-      {/* 420px 이상에서만 Navbar를 보여줌 */}
-      <div className="hidden xs:block">
-        <Navbar />
-      </div>
-
-      {/* 420px 이하에서만 Footer를 보여줌 */}
-      <div className="block xs:hidden">
-        <Footer />
-      </div>
-
-      {/* 메인 콘텐츠 영역 */}
-      <main className="flex justify-center">{children}</main>
-    </div>
+    <MapLayout>
+      <main>{children}</main>
+    </MapLayout>
   );
 };
